Move convertToModel to InstrumentationConfigUtil
