+++ sdk-extensions/incubator/src/main/java/io/opentelemetry/sdk/extension/incubator/fileconfig/YamlDeclarativeConfigProperties.java
@@ -11,7 +11,6 @@ import static java.util.stream.Collectors.toList;
 import io.opentelemetry.api.incubator.config.DeclarativeConfigException;
 import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.internal.ComponentLoader;
-import io.opentelemetry.sdk.extension.incubator.fileconfig.internal.model.OpenTelemetryConfigurationModel;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -67,10 +66,10 @@ public final class YamlDeclarativeConfigProperties implements DeclarativeConfigP
    * com.fasterxml.jackson.databind.ObjectMapper}), and have values which are scalars, lists of
    * scalars, lists of maps, and maps.
    *
-   * @see DeclarativeConfiguration#toConfigProperties(OpenTelemetryConfigurationModel)
+   * @see DeclarativeConfiguration#toConfigProperties(Object)
    */
   @SuppressWarnings("unchecked")
-  static YamlDeclarativeConfigProperties create(
+  public static YamlDeclarativeConfigProperties create(
       Map<String, Object> properties, ComponentLoader componentLoader) {
     Map<String, Object> simpleEntries = new LinkedHashMap<>();
     Map<String, List<YamlDeclarativeConfigProperties>> listEntries = new LinkedHashMap<>();
